+++ src/test/java/org/chocosolver/solver/search/loop/monitors/CPProfilerTest.java
@@ -10,6 +10,7 @@ package org.chocosolver.solver.search.loop.monitors;
 
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.search.loop.lns.neighbors.RandomNeighborhood;
+import org.chocosolver.solver.trace.CPProfiler;
 import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.util.ProblemMaker;
 import org.testng.annotations.Test;